% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols-to-ebd.R
\name{cols_to_ebd}
\alias{cols_to_ebd}
\title{Convert columns in eBird data to EBD style}
\usage{
cols_to_ebd(data, type = "mydata")
}
\arguments{
\item{data}{eBird data object}

\item{type}{character; "mydata" if personal My Data download}
}
\value{
An updated data frame with new column names, and some columns changed:
\code{PROTOCOL.TYPE} and \code{HAS.MEDIA} (from \verb{ML Catalog Numbers}) values converted to EBD style
}
\description{
The column names in some datasets such as My Data are not ideal: they contain
white spaces and follow complex capitalisation. In such cases, \code{cols_to_ebd()}
can be used to standardise the column names to match the default style as in
\code{read.ebd()}, i.e., fully uppercase and with periods replacing spaces.
}
