% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ebd_admin_units.R
\docType{data}
\name{ebd_admin_units}
\alias{ebd_admin_units}
\title{Administrative units of the Indian subcontinent}
\format{
A tibble with 1,135 observations
and 6 variables (names and 2-letter ISO codes).
}
\usage{
ebd_admin_units
}
\description{
Names and codes of administrative units (three levels) used by eBird in their datasets
for the eight countries of the Indian subcontinent.

There is often a mismatch in information on administrative units in different sources,
so this data can be useful for any work involving eBird data, particularly in cases
with low data where the downloaded dataset might not have observations from all
relevant units or subunits.

This data is obtained from the API and should therefore stay up to date with each package
release. However, it may be outdated if changes occur before the package is updated. In
such cases, one can use \code{rebird::ebirdsubregionlist()} manually (section See Also).
}
\examples{
  ebd_admin_units
  # get data for only India
  dplyr::filter(ebd_admin_units, COUNTRY.CODE == "IN")

}
\seealso{
The script to update the data in \code{skimmr}:
\url{https://github.com/rikudoukarthik/skimmr/blob/main/data-raw/ebd_admin_units.R}
The \code{rebird} package for interacting with eBird webservices: \url{https://docs.ropensci.org/rebird/}
}
\keyword{datasets}
