% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_rel_param.R
\name{ebird_rel_param}
\alias{ebird_rel_param}
\title{EBD parameters}
\usage{
ebird_rel_param(
  date_real = NULL,
  date_currel = NULL,
  date_prevrel = NULL,
  extra = FALSE
)
}
\arguments{
\item{date_real}{a date string}

\item{date_currel}{a date string}

\item{date_prevrel}{a date string}

\item{extra}{whether extra parameters are needed}
}
\value{
multiple date objects
}
\description{
Get date parameters based on current or specified date
}
\examples{
ebird_rel_param()
ebird_rel_param(extra = TRUE)
ebird_rel_param("2023-10-01")
}
