% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-eligible.R
\name{filter_eligible}
\alias{filter_eligible}
\title{Filter eBird data for eligible checklists}
\usage{
filter_eligible(
  data,
  standard = "bci",
  complete = NULL,
  duration = NULL,
  counts = NULL
)
}
\arguments{
\item{data}{eBird data object}

\item{standard}{character; a standard eligibility criterion (default "bci", or "ebird")}

\item{complete}{logical; whether or not to filter for complete checklists (default NULL)}

\item{duration}{numerical; whether or not to filter for a specified checklist duration (>=; default NULL)}

\item{counts}{logical; whether or not to filter for true counts (i.e., remove lists reporting Xs; default NULL)}
}
\value{
An updated eBird data object filtered as specified
}
\description{
For various usecases, it is required to filter the eBird dataset and use only "eligible"
checklists. An example is the monthly eBird challenges run by various collectives across
the world. As such, definitions of "eligibility" tend to vary, but there are two popularly
used definitions which can be accessed via the \code{standard} argument. On the other hand,
it is also possible to create a custom eligibility filter, by setting \code{standard = NULL}
and providing input to the other arguments.
}
\seealso{
Eligibility criteria for
\href{https://support.ebird.org/en/support/solutions/articles/48000948757-ebird-faqs#anchorEoMchallenge}{eBird monthly challenges}
(standard = "ebird") and for \href{https://birdcount.in/aug24-challenge/}{BCI monthly challenges}
(standard = "bci)
}
