% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param.R
\name{get_param}
\alias{get_param}
\title{EBD parameters}
\usage{
get_param(
  date_real = NULL,
  date_currel = NULL,
  date_prevrel = NULL,
  extra = FALSE
)
}
\arguments{
\item{date_real}{a date string}

\item{date_currel}{a date string}

\item{date_prevrel}{a date string}

\item{extra}{whether extra parameters are needed}
}
\value{
multiple date objects
}
\description{
Get date parameters based on current or specified date
}
\examples{
get_param()
get_param(extra = TRUE)
get_param("2023-10-01")
}
