% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-india_checklist.R
\docType{data}
\name{india_checklist}
\alias{india_checklist}
\title{Checklist of birds of India}
\format{
A tibble with 1,368 observations
and 25 variables. Please see original \code{.xlsx} file
(Source below) for descriptions of each variable.
}
\source{
\url{https://indianbirds.in/india/}
}
\usage{
india_checklist
}
\description{
This is the comprehensive checklist of birds of India, with its own taxonomy. The
current version of the checklist contains 1368
species.

Please cite this online checklist as:

Praveen J., & Jayapal, R., 2024. Checklist of the birds of India (v8.2). Website: http://www.indianbirds.in/india/ (Date of publication: 20 August 2024)
}
\examples{
  india_checklist

}
\seealso{
The script to update the latest version of the checklist in \code{skimmr}:
\url{https://github.com/rikudoukarthik/skimmr/blob/main/data-raw/india_checklist.R}
}
\keyword{datasets}
