% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimmr-package.R
\docType{package}
\name{skimmr-package}
\alias{skimmr}
\alias{skimmr-package}
\title{skimmr: Working with eBird data made easy}
\description{
Functions for various data sorting, processing and analysis tasks of citizen science practitioners using eBird data.
}
\author{
\strong{Maintainer}: Karthik Thrikkadeeri \email{rikudoukarthik@gmail.com} (\href{https://orcid.org/0000-0003-2863-5879}{ORCID})

Other contributors:
\itemize{
  \item Ashwin Viswanathan \email{ashwinv2005@gmail.com} (\href{https://orcid.org/0000-0002-1605-0614}{ORCID}) [contributor]
}

}
\keyword{internal}
