% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-soib.R
\docType{data}
\name{soib}
\alias{soib}
\title{State of India's Birds (SoIB) datasheet}
\format{
A tibble with 1357 observations
and 51 variables. Please see original \code{.xlsx} file
(Source below) for descriptions of each variable.
}
\source{
\url{https://zenodo.org/records/11124589?}
}
\usage{
soib
}
\description{
This is the latest State of India's Birds (SoIB) datasheet at the national level. The
current version of the datasheet contains 1356
species.

Please cite this datasheet as:

SoIB. (2023). State of India's Birds, 2023: Range, trends, and conservation status (p. 119). The SoIB Partnership. https://doi.org/10.5281/zenodo.11124590
}
\examples{
  soib

}
\seealso{
The script to update the latest version of the datasheet in \code{skimmr}:
\url{https://github.com/rikudoukarthik/skimmr/blob/main/data-raw/soib.R}
}
\keyword{datasets}
