% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soib_skim.R
\name{soib_skim}
\alias{soib_skim}
\title{Get State of India's Birds (SoIB) datasheet for specific mask}
\usage{
soib_skim(mask)
}
\arguments{
\item{mask}{character; a valid mask name}
}
\value{
A tibble. Please find info on descriptions of each variable in the
source listed in the documentation for \link{soib}.
}
\description{
Whereas \code{soib} is the latest State of India's Birds (SoIB) datasheet at the national level,
\code{soib_skim()} lets you obtain the SoIB datasheet for a specific subnational
(or national, with \code{mask == "India"}) mask of your choice, such as habitat masks
or individual states.
}
\examples{
soib_skim("ONEs") # get SoIB data for Open Natural Ecosystems (ONEs)
}
