% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfn.R
\name{testfn}
\alias{testfn}
\title{Test and answer to life}
\usage{
testfn(x)
}
\arguments{
\item{x}{any input}
}
\value{
A character test string, and the integer 42
}
\description{
Test and answer to life
}
\examples{
x <- "Nihilism"
testfn(x)
}
