% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_ebd.R
\name{unzip_ebd}
\alias{unzip_ebd}
\title{Unzip eBird Basic Dataset (EBD) file}
\usage{
unzip_ebd(
  dataset_str = "ebd_IN_unv_smp_rel",
  unzip_which = c("ebd", "sed"),
  dir_name = NULL
)
}
\arguments{
\item{dataset_str}{character; the string in the file name of the data download, ending with \verb{_rel}}

\item{unzip_which}{character; vector listing which of "ebd", "sed" and "unv" to unzip}

\item{dir_name}{character; specify subfolder name of zip file (if exists), without
the trailing slash}
}
\value{
Does not return output. Specified files are unzipped and saved in the local directory.
}
\description{
Unzips an EBD data download file, and saves any of three selected data files: main dataset
of vetted observation records, sampling event data (SED), and unvetted records.

\code{dataset_str} varies with options selected when downloading the data; it may contain
district-level admin. codes, or species codes, or tags indicating whether or not SED
and unvetted data are included.
}
