% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ebd.R
\name{read.ebd}
\alias{read.ebd}
\title{EBD Input}
\usage{
read.ebd(ebd_path, cols_sel = "all", cols_print_only = FALSE)
}
\arguments{
\item{ebd_path}{character; the path to the downloaded EBD .txt file}

\item{cols_sel}{character; vector of column names to be imported from the dataset}

\item{cols_print_only}{logical; whether or not to only print the full set of column
names}
}
\value{
A data frame
}
\description{
Reads a .txt EBD file and creates a data frame from it, with cases corresponding to
lines and variables to fields in the file. This function is a wrapper around
\code{utils::read.delim()}, which is considerably faster than the \code{readr::read_delim()} used in
\code{auk::read_ebd()}. Moreover, unlike the latter which uses snake case for column names,
this function uses uppercase with period separators.
}
\details{
Since EBD contains several columns, which may not all be required for a given usecase,
\code{cols_sel} can be used to import only a subset of the columns. To see the list of all
columns names to choose from, run \verb{read.ebd(ebd_path, cols_print_only = TRUE}.
}
\examples{
# to see list of column names before choosing
test1 <- c(SAMPLING.EVENT.IDENTIFIER = "S0000001", COMMON.NAME = "Indian Peafowl")
tf <- tempfile()
writeLines(test1, tf)
read.ebd(tf, cols_print_only = TRUE)

# select columns and import data
read.ebd(tf, cols_sel = c("SAMPLING.EVENT.IDENTIFIER", "COMMON.NAME"))
}
